ai/control/agent-ops/assets/wxo-agents/requirements.txt
ibm-watsonx-orchestrate[agentops]>=2.5.1,<2.9.0
ibm-watsonx-orchestrate-evaluation-framework>=1.2.7,<2.0.0
langfuse<4
pandas>=2.1.0
python-dotenv>=1.0.0
requests>=2.28.0
